Gun Controls
Implemented a basic UI show where the gun will fire a repulsion or attraction well, the power of the well and the cycle of the wells which removes the first well activated when there a 4th well is needed.
Added a capsule on the gun to indicate the power of the gun, the capsule scales depending on the power adjusted by the user.
When the user holds the fire button a preview of where a gravity well will land and what it will affect. A transparent sphere shows the radius and the colour show the polarity.
The gravity well is now launched on release of the fire button.
Picking up objects is now easer as it is on button down instead of the axis system.